Add files via upload
Updated Hangman game with Player class, which also handles the leaderboard. Player can now play multiple rounds of Hangman and accumulate points. Once the player exits, leaderboard is updated and Player is reset. Unable to sort at this time as only solution found required use of auto variables, but will continue to research. Also need to complete the output to a leaderboard.txt file, but wanted this piece reviewed first.